mcp_hotel_server/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
